system_focus_manager/requirements.txt
# Requirements for System Focus Manager
# Copyright © 2025 Manuela Riascos Hurtado
# Author: manhurta54@gmail.com

# GUI Framework
PySide6>=6.6.0
PySide6-Essentials>=6.6.0

# Process Management
psutil>=5.9.0

# Browser Control
requests>=2.31.0
websocket-client>=1.6.0

# Keyboard Control
keyboard>=0.13.5

# System Tray
pystray>=0.19.0
Pillow>=10.0.0

# Windows-specific
pywin32>=306; platform_system == "Windows"
